requirements.txt
openai-whisper
sounddevice
pyttsx3
pvporcupine
pyaudio
streamlit
pandas
openai


